backend/requirements.txt
# =============================================================================
# 📦 Dépendances Python Backend
# =============================================================================

# =============================================================================
# 🌐 FRAMEWORK WEB
# =============================================================================

# FastAPI (recommandé pour APIs REST modernes)
fastapi==0.104.1
uvicorn==0.24.0

# =============================================================================
# 🗃️ CONNECTEURS BASE DE DONNÉES
# =============================================================================

# PostgreSQL
psycopg2-binary==2.9.9

# =============================================================================
# 🔧 OUTILS ET UTILITAIRES
# =============================================================================

# Requêtes HTTP pour health checks
requests==2.31.0

# Variables d'environnement
python-dotenv==1.0.0

# Validation de données
pydantic==2.5.0

# =============================================================================
# 🏥 HEALTH CHECKS ET MONITORING
# =============================================================================

# Pour les health checks HTTP
httpx==0.25.2